Change point struct to use floats rather than 16bit signed integers. This allows a simplification of the transformation API

+++ low_poly_engine/EnginePrimitives.h
@@ -12,9 +12,9 @@
 #include <Metal/Metal.h>
 
 typedef struct LPPoint{
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    float x;
+    float y;
+    float z;
 } LPPoint;
 
 typedef struct LPTriangle{
